% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuro_icds_10}
\alias{neuro_icds_10}
\title{ICD-10 codes of neurological conditions, grouped in to
CNS (Central Nervous System) and PNS (Peripheral Nervous System).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 5 columns.
}
\usage{
neuro_icds_10
}
\description{
ICD-10 codes of neurological conditions, grouped in to
CNS (Central Nervous System) and PNS (Peripheral Nervous System).
}
\keyword{datasets}
