% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuro_icds_9}
\alias{neuro_icds_9}
\title{ICD-9 codes of neurological conditions, grouped in to
CNS (Central Nervous System) and PNS (Peripheral Nervous System).}
\format{
An object of class \code{data.frame} with 29 rows and 5 columns.
}
\usage{
neuro_icds_9
}
\description{
ICD-9 codes of neurological conditions, grouped in to
CNS (Central Nervous System) and PNS (Peripheral Nervous System).
}
\keyword{datasets}
