% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run analysis}
\usage{
runAnalysis()
}
\value{
NULL. Result files are written out to the `getProjectOutputDirectory()` directory.
}
\description{
Run analysis
}
\keyword{4CE}
