% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_params}
\alias{site_params}
\title{Parameters of site-specific data.
5 columns: siteid, mask_thres, blur_abs, icd_version, include_race.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 5 columns.
}
\usage{
site_params
}
\description{
Parameters of site-specific data.
5 columns: siteid, mask_thres, blur_abs, icd_version, include_race.
}
\keyword{datasets}
