% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateAnalysis.R
\name{validateAnalysis}
\alias{validateAnalysis}
\title{Validates the results of the analytic workflow for the Neuro project}
\usage{
validateAnalysis()
}
\description{
Validates the results of the analytic workflow for the Neuro project
}
\keyword{4CE}
