% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run analysis}
\usage{
runAnalysis(mask_thres, blur_abs, include_race = TRUE, data_dir = "Input")
}
\arguments{
\item{mask_thres}{Obfuscation small count mask threshold (e.g. 10)}

\item{blur_abs}{Absolute max of obfuscation blurring range}

\item{include_race}{Boolean. Whether race should be included
in the regression model.}

\item{data_dir}{Optional. Directory where datasets are located.
Defaults to 'Input'.}
}
\value{
NULL. Result files are written out to the `getProjectOutputDirectory()` directory.
}
\description{
Run analysis
}
\keyword{4CE}
