% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run analysis}
\usage{
runAnalysis()
}
\arguments{
\item{mask_thres}{Obfuscation small count mask threshold (e.g. 10)}

\item{blur_abs}{Absolute max of obfuscation blurring range}

\item{icd_version}{Version of ICD code the site uses. Must be EITHER 9 or 10.}

\item{include_race}{Boolean. Whether race should be included
in the regression model.}
}
\value{
NULL. Result files are written out to the `getProjectOutputDirectory()` directory.
}
\description{
Run analysis
}
\keyword{4CE}
